int -> short for category list position

+++ app/src/main/java/com/silver/dan/deals/ProductArrayAdapter.java
@@ -37,10 +37,10 @@ public class ProductArrayAdapter extends RecyclerView.Adapter<ProductArrayAdapte
     }
 
     //count how many products for each brand
-    public HashMap<String, Integer> getBrandCounts() {
-        HashMap<String, Integer> brandCounts = new HashMap<>();
+    public HashMap<String, Short> getBrandCounts() {
+        HashMap<String, Short> brandCounts = new HashMap<>();
         for (Product p : allProducts()) {
-            Integer count = 1;
+            Short count = 1;
             if (brandCounts.containsKey(p.brand)) {
                 count = brandCounts.get(p.brand);
                 count++;
